requirements.txt
python-dotenv
Flask
slack-bolt
slack-sdk
openai
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai

